requirements.txt
biopython
alabaster==0.7.12
appnope==0.1.0
argh==0.26.2
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
biopython==1.74
bleach==3.1.4
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
commonmark==0.9.1
configparser==4.0.2
coverage==4.5.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docker-pycreds==0.4.0
docutils==0.15.2
entrypoints==0.3
freemol==1.158
future==0.18.2
gitdb2==2.0.6
GitPython==3.0.4
gql==0.2.0
graphql-core==1.1
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.2.0
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.17.2
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pathtools==0.1.2
pdb2pqr-minimal==0.0.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
powerline-status==2.7
ProDy==1.10.10
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.9
-e git+git@github.com:jonathanking/protein-transformer.git@f67e0ecad6336269a27324f9215dba1762550a0e#egg=protein_transformer
psutil>=5.6.6
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
pymol==2.4.0a0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.2.2
pytest-cov==2.8.1
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.7.2
QtPy==1.9.0
recommonmark==0.6.0
requests==2.22.0
Rx==1.6.1
scipy==1.3.1
Send2Trash==1.5.0
sentry-sdk==0.13.1
shortuuid==0.5.0
-e git+git@github.com:jonathanking/sidechainnet.git@b5a9ff7bc0acb62244c5f902587f1870f2e9df76#egg=sidechainnet
six==1.12.0
smmap2==2.0.5
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
subprocess32==3.5.4
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.25.6
wandb==0.8.31
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
